Move Action types to its own folder
